Moved fetchPreProposeAddress function to proposal module adapter so proposal modules can implement pre-propose however they choose.
