+++ apps/dapp/pages/inbox.tsx
@@ -11,14 +11,15 @@ import {
   ProposalLine,
   ProposalLineProps,
   SuspenseLoader,
+  pinnedDaoDropdownInfosSelector,
 } from '@dao-dao/common'
 import { serverSideTranslations } from '@dao-dao/i18n/serverSideTranslations'
-import { pinnedDaoDropdownInfosSelector } from '@dao-dao/state'
 import {
   DaoWithProposals,
   Inbox,
   PageLoader,
   ProfileDisconnectedCard,
+  useDaoInfoContext,
 } from '@dao-dao/ui'
 import { SITE_URL } from '@dao-dao/utils'
 
@@ -26,6 +27,7 @@ import { ProfileHomeCard, useDAppContext } from '@/components'
 
 const InnerInbox = () => {
   const { connected } = useWallet()
+  const { chainId } = useDaoInfoContext()
   const pinnedDaoDropdownInfos = useRecoilValue(pinnedDaoDropdownInfosSelector)
 
   const {
@@ -50,9 +52,10 @@ const InnerInbox = () => {
           dao: daoDropdownInfo,
           proposals: openUnvotedProposals.map(
             ({ proposalModule: { prefix }, proposalNumber }) => ({
+              chainId,
+              coreAddress,
               proposalId: `${prefix}${proposalNumber}`,
               proposalModules,
-              coreAddress,
               proposalViewUrl: `/dao/${coreAddress}/proposals/${prefix}${proposalNumber}`,
             })
           ),
