+++ packages/proposal-module-adapter/adapters/CwdProposalSingle/functions/index.ts
@@ -1,1 +1,2 @@
+export * from './fetchPreProposeAddress'
 export * from './makeGetProposalInfo'
