+++ packages/state/index.ts
@@ -2,4 +2,3 @@ export * from './clients'
 export * from './hooks'
 export * from './recoil'
 export * from './subquery'
-export * from './utils'
