+++ packages/state/recoil/selectors/dao/index.ts
@@ -1,3 +1,2 @@
 export * from './info'
-export * from './pinned'
 export * from './treasury'
