+++ packages/state/recoil/selectors/proposal.ts
@@ -1,11 +1,8 @@
 import { selectorFamily } from 'recoil'
 
 import { WithChainId } from '@dao-dao/tstypes'
-import { ProposalModule } from '@dao-dao/tstypes/dao'
 
-import { fetchProposalModules } from '../../utils'
 import { cosmWasmClientForChainSelector } from './chain'
-import { contractVersionSelector } from './contract'
 
 export const proposalExecutionTXHashSelector = selectorFamily<
   string | undefined,
@@ -30,25 +27,5 @@ export const proposalExecutionTXHashSelector = selectorFamily<
       }
 
       return events?.[0]?.hash
-    },
-})
-
-export const cwCoreProposalModulesSelector = selectorFamily<
-  ProposalModule[],
-  WithChainId<{ coreAddress: string }>
->({
-  key: 'cwCoreProposalModules',
-  get:
-    ({ coreAddress, chainId }) =>
-    async ({ get }) => {
-      const client = get(cosmWasmClientForChainSelector(chainId))
-      const coreVersion = get(
-        contractVersionSelector({
-          contractAddress: coreAddress,
-          chainId,
-        })
-      )
-
-      return await fetchProposalModules(client, coreAddress, coreVersion)
     },
 })
